+++ apps/web/src/lib/config/index.ts
@@ -3,3 +3,4 @@
  */
 
 export * from './inputFields';
+export * from './wizard';